+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Created automation for creating instances in the chatwoot bot with the command '#inbox_whatsapp:<INSTANCE_NAME>'
 * Change Baileys version to: 6.4.0
 * Send contact in chatwoot
+* Send contact array in chatwoot
 
 ### Fixed